+++ src/socket/socket.io.js
@@ -239,13 +239,6 @@ class SocketIOServer extends SocketServer {
     }
   }
 
-  respond(socket, statusCode, body) {
-    super.respond(socket, statusCode, body);
-    if (statusCode >= 400 && socket.request?._next) {
-      socket.request?._next(new Error(body));
-    }
-  }
-
   close(socket) {
     if (socket) {
       socket.disconnect(true);