+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- tbd
+- Improve redis connection check
+- Improve connection error handling (e.g. unauthorized) 
 
 ## Version 1.5.0 - 2025-01-08